+++ CHANGELOG.md
@@ -14,6 +14,11 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - The `Enum` derive now supports fallback variants
 - Added `Operation::new` to allow `Operation` to be used in tests.
 
+### Removed
+
+- Removed the deprecated `FragmentArguments` derive.
+- Removed the deprecated `arugment_struct` attribute.
+
 ## v3.0.0-beta.3 - 2023-05-13
 
 ### New Features